+++ src/types/Config.tsx
@@ -88,6 +88,7 @@ import { OSPConsent } from 'components/pages/OSPConsent'
 import CompanySubscriptions from 'components/pages/CompanySubscriptions'
 import CompanySubscriptionDetail from 'components/pages/CompanySubscriptions/CompanySubscriptionDetail'
 import CompanyData from 'components/pages/CompanyData'
+import OnboardingServiceProvider from 'components/pages/OnboardingServiceProvider'
 
 /**
  * ALL_PAGES
@@ -595,6 +596,10 @@ export const ALL_PAGES: IPage[] = [
     role: ROLES.MY_ORGANIZATION_VIEW,
     element: <CompanyData />,
   },
+  {
+    name: PAGES.ONBOARDING_SERVICE_PROVIDER,
+    element: <OnboardingServiceProvider />,
+  },
 ]
 
 export const ALL_OVERLAYS: IOverlay[] = [
@@ -828,6 +833,7 @@ export const userMenuFull = [
   PAGES.COMPANY_CERTIFICATE,
   PAGES.COMPANY_WALLET,
   PAGES.COMPANY_DATA,
+  PAGES.ONBOARDING_SERVICE_PROVIDER,
   PAGES.LOGOUT,
 ]